requirements.txt
django
gunicorn
django-heroku
djangorestframework
django-rest-auth
django-allauth
django-cors-headers